Braking down basic_blocks into 3 separate modules
