+++ lib_dsp/mult_dsp/__init__.py
@@ -1,0 +1,1 @@
+from .mult_dsp.design.mult_dsp import *
